Implemented DataContext
Implemented the DataContext and connected the DB to the Server project. I need to go to bed because I am falling asleep in the chair, but I will continue tomorrow because right now I literally cannot think.
P.S. Wow, are you still reading this?! Well you did it, this is the end of this long description.
THE END